have pagination component calculate results count
