+++ docs/answers-react-components.pagination.md
@@ -1,0 +1,24 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Pagination](./answers-react-components.pagination.md)
+
+## Pagination() function
+
+Renders a component that divide a series of results into chunks across multiple pages and enable user to navigate between those pages.
+
+<b>Signature:</b>
+
+```typescript
+export declare function Pagination(props: PaginationProps): JSX.Element | null;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  props | [PaginationProps](./answers-react-components.paginationprops.md) |  |
+
+<b>Returns:</b>
+
+JSX.Element \| null
+
