+++ docs/answers-react-components.paginationcssclasses.md
@@ -20,6 +20,7 @@ export interface PaginationCssClasses
 |  [label?](./answers-react-components.paginationcssclasses.label.md) | string | <i>(Optional)</i> |
 |  [labelContainer?](./answers-react-components.paginationcssclasses.labelcontainer.md) | string | <i>(Optional)</i> |
 |  [leftIconContainer?](./answers-react-components.paginationcssclasses.lefticoncontainer.md) | string | <i>(Optional)</i> |
+|  [paginationContainer\_\_\_loading?](./answers-react-components.paginationcssclasses.paginationcontainer___loading.md) | string | <i>(Optional)</i> |
 |  [paginationContainer?](./answers-react-components.paginationcssclasses.paginationcontainer.md) | string | <i>(Optional)</i> |
 |  [rightIconContainer?](./answers-react-components.paginationcssclasses.righticoncontainer.md) | string | <i>(Optional)</i> |
 |  [selectedLabel?](./answers-react-components.paginationcssclasses.selectedlabel.md) | string | <i>(Optional)</i> |
