+++ docs/answers-react-components.paginationcssclasses.paginationcontainer___loading.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [PaginationCssClasses](./answers-react-components.paginationcssclasses.md) &gt; [paginationContainer\_\_\_loading](./answers-react-components.paginationcssclasses.paginationcontainer___loading.md)
+
+## PaginationCssClasses.paginationContainer\_\_\_loading property
+
+<b>Signature:</b>
+
+```typescript
+paginationContainer___loading?: string;
+```
