+++ docs/answers-react-components.paginationprops.csscompositionmethod.md
@@ -1,0 +1,18 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [PaginationProps](./answers-react-components.paginationprops.md) &gt; [cssCompositionMethod](./answers-react-components.paginationprops.csscompositionmethod.md)
+
+## PaginationProps.cssCompositionMethod property
+
+The method of combining a component's built-in CSS classes with custom CSS classes.
+
+<b>Signature:</b>
+
+```typescript
+cssCompositionMethod?: CompositionMethod;
+```
+
+## Remarks
+
+Merge keeps the component's built-in classes and adds the custom classes to them (default). Replace ignore all of the component’s built-in classes and only uses the custom classes. Assign keeps the component's built-in classes, however custom classes will completely override their associated built-in classes.
+
