+++ docs/answers-react-components.paginationprops.customcssclasses.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [PaginationProps](./answers-react-components.paginationprops.md) &gt; [customCssClasses](./answers-react-components.paginationprops.customcssclasses.md)
+
+## PaginationProps.customCssClasses property
+
+CSS classes for customizing the component styling.
+
+<b>Signature:</b>
+
+```typescript
+customCssClasses?: PaginationCssClasses;
+```
