+++ docs/answers-react-components.paginationprops.md
@@ -1,0 +1,22 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [PaginationProps](./answers-react-components.paginationprops.md)
+
+## PaginationProps interface
+
+Props for [Pagination()](./answers-react-components.pagination.md) component
+
+<b>Signature:</b>
+
+```typescript
+export interface PaginationProps 
+```
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [cssCompositionMethod?](./answers-react-components.paginationprops.csscompositionmethod.md) | [CompositionMethod](./answers-react-components.compositionmethod.md) | <i>(Optional)</i> The method of combining a component's built-in CSS classes with custom CSS classes. |
+|  [customCssClasses?](./answers-react-components.paginationprops.customcssclasses.md) | [PaginationCssClasses](./answers-react-components.paginationcssclasses.md) | <i>(Optional)</i> CSS classes for customizing the component styling. |
+|  [paginateAllOnNoResults](./answers-react-components.paginationprops.paginateallonnoresults.md) | boolean | Whether or not to paginate based on the total results count of the vertical when there are none returned from the search. Defaults to false. |
+
