+++ docs/answers-react-components.paginationprops.paginateallonnoresults.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [PaginationProps](./answers-react-components.paginationprops.md) &gt; [paginateAllOnNoResults](./answers-react-components.paginationprops.paginateallonnoresults.md)
+
+## PaginationProps.paginateAllOnNoResults property
+
+Whether or not to paginate based on the total results count of the vertical when there are none returned from the search. Defaults to false.
+
+<b>Signature:</b>
+
+```typescript
+paginateAllOnNoResults: boolean;
+```
